requirements.txt
# Core runtime dependencies for the cleaned public release.

python-dotenv>=0.19.0
pyyaml>=6.0
numpy>=1.21.0
scipy>=1.7.0
requests>=2.28.0

# LLM / NLP
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.25.0
sentence-transformers>=2.2.0

# Code analysis / parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Data / utilities
pydantic>=1.10.0
dataclasses-json>=0.5.0
tqdm>=4.64.0
loguru>=0.6.0
